Make `tstpnt2.py` pytest-ready
